pyproject.toml
[tool.poetry]
name = "import_api"
version = "0.1.0"
description = ""
authors = ["abishekk92 <bhat.abishek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
newspaper3k = "^0.2.8"
markdownify = "^0.11.2"
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
uvicorn = "^0.18.1"
magnum = "^14.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
uvicorn = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
